backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pdf2docx==0.5.6
werkzeug==3.0.1
zhipuai==2.0.1
python-dotenv==1.0.0
